Constant Element References
Moved over all ID element references from the window global-based kind, to constant query selector ones. This is to help narrow down the types gap that each module appears to need access too. This will help me with further breaking up the codebase into distinct modules, components, and imports/exports. It's hard for both me and TypeScript to decypher what each module relies on when it is getting references to things outside of it's scope not using static import statements.

Discovered a Vite build caching issue, where while running `npm run dev`, the main `app.ts` script would run twice, causing all kinds of errors on the initial page load.

Not totally sure if it's what fixed it (The issue did go away after I tried it though), but looks like you can rebuild the dependency build cache using the `--force` flag, neat! I'm guessing the Vite server was trying to load two versions, one from the cache, and one that was live-loaded for some reason? It was weird though, I would add an addition single `console.log()` at the top of the script, and that would definitely run twice. So it's not like it was an old version in the cache that was loading twice, it was the same new fresh build that was loading twice. It would successfully run the updates that I added to the code, just twice XD

Should I say twice, twice as many times?

https://github.com/vitejs/vite/issues/10986#issuecomment-1321981632