Rename Menu Fix + Editor Component
Finally got the Rename Editor file menu item working again! Turns out it was seriously just a typo from the API name changes I did a few months ago... uugh. This is part of the logic that's adding into my head in regards to moving towards React or something like that. I want to get TypeScript/JavaScript checking in HTML too, but that doesn't seem to be something that works very well unfortunately.

Are there any React-ish frameworks out there that simply do things using Web Components as a back-end for the library, vanilla DOM events, and can use JSX/TypeScript? That sounds pretty much ideally like what I'm looking for. I really like the look of the React workflow, but I haven't (because of my inexperience I think, hehe. This will change, I strongly think. Just gotta dive deeper in the learning process) been fond of the React ecosystem quite yet, since it requires more of a build step and bundling process. I don't want to include extra bloat that I don't need. I simply want a nice API abstraction on top of Web Components, HTML, and TypeScript, which will nicely go down to vanilla ESM for the browser. With my early knowledge of everything, I think that's essentially Svelte? Not sure about that take, but I think that sounds nearly close to that.

This Vite dev runner bug is killing me, mentioned it in the last commit. Anyone have any ideas as to what's causing this? It's when I update the source for my app. I have a hunch it's because of my recursive-ish module structures maybe. It started happening once I abstracted the DOM code references into it's own module.

Made the Editor component an extension of the NumText component itself! This simplifies the DOM a bit, as the Editor object itself is just essentially the NumText element. But since this is a Web Component, it is the NumText element, plus everything else that the Editor component requires functionality of. Super neat!

Noticed that I forgot I added a `setTimeout()` handler to MenuDrop's onclick handlers already, so I could remove it from STE's Rename Editor file menu item.

This was pretty much done a few days ago, but I got frustrated with the Rename menu option still not being able to be fixed, so I had to come back to it later. I'm relieved and both annoyed that it was just a typo this whole time, as I looked over that part of the HTML a few times now, and that I had missed renaming that one when I did the migration a few months ago. I'm getting inspired towards learning how to build things using different frameworks, TypeScript and ESM works nice for vanilla projects, but it is starting to get a bit hairy, even with type checking, as you can't rely on it in all places, since it doesn't span the whole codebase to include things like HTML. I'm not sure if going full JS is the answer, as it seems to be frowned upon from some other programmers because of the cruft it adds to the site build (I feel the concern for that too), but maybe that is just the way to go as of now in terms of getting a nice developer experience set up. Maybe that's what my next goal is! For a while I have been trying to get a really nice simple production build for most of my projects, and that's been the end goal, sometimes sacrificing DX. Maybe my next exploration in the learning cave can be to try getting the nicest dev experience set up :)