Module Cross-Pollination
Figured out what was causing the module loading issues in vite dev, and the Editor loading problems!

The Vite part was because I was both importing and exporting values in and out of the main `app.ts` entry point, and while the browser seems to be able to handle this (I think it's part of the spec too), it doesn't look like Vite can do this with the live server. Moving the main module's exports into a secondary module fixed the resolution errors (I think it was causing the main module to load multiple times, causing errors for the code in the app that didn't expect the app to be run twice, lol).

And for the Editor part, changing the text value to the actual internal textarea's value property is the way to go, since `NumTextElement` uses a getter that enables undo-redo, causing the weird auto-created behavior for when opening files using File Handling. And also looks like `await`-ing the Service Worker setup was delaying the auto-created setup, which was unintentionally setting an extra time delay from when it should have been run.

Now I see how I can refactor the code better, I think the life cycle of the app is the main unexpected/undocumented part of the current stack, with a mix of async and sync code, things aren't specifically running in absolute execution orders. Now that things are type-safe, I think I have more safety in reliably moving sections of the app around, and straightening out the order in the lifecycle of how the app flows time-wise.