File Handle-ing + Johnny Deppendencies
Updated project dependencies to their latest versions, and started tidying up some of the loose ends of the codebase that haven't moved over to use my newer implementations yet. In some cases, I have new implementations setup to make things more component-based, and there's still some parts that are defined in more than once place, from before they were unified into a single implementation. I want to do that with more parts of the codebase, like for opening and saving files, and with drag and drop. I like what I've recently done in that regard, over in Dovetail. I want STE to have a similar abstraction pattern, where main ideas are implemented only in one place, and you just have to connect certain pipes together to get it working, rather than rewriting a whole new plumbing system right adjacent to the pre-existing one (that was what I unintentionally tended to do with STE's development). Having TypeScript and module experience has helped me see these kinds of flows a bit better, to where you build more generic implementations that do a single thing, and you can hook multiple of those together to build a new thing altogether, rather than implementing it from the groundUp again. This also helps a bunch when you need to change or upgrade how one of those building blocks work, as it will just be reflected right to the other ones, you don't have to make sure they are in-sync or the same, because instead they just use the same base-implementation. This might kind of sound like OOP, but I think I try to do things more Functionally, where you pipe values in and out of functions that you build, which do certain things. And for tasks that require on-going state, I think I like to use Classes for those, with publicly accessible methods and properties which managed the internal state of that Object. Right now, STE uses that model, but the state is external to the Object, stored in the global scoped STE object. Not sure what that model's called, where it's like a centralized state holder or something? I see why I did that, and I do think it's still reasonable to build something like that, if you task is simple enough to follow that pattern. But when state tends to be local to instances of each component, then managing the state from a place outside of the component seems to make less sense the further you have to add cords to plug the component and the state machine together. You have to do checks in between the two a bunch, and it kind of makes it hard to tell what is using what, and where.

For example, right now the file handles for each editor are stored in a global map, and the index is by each Editor's identifier, which is then queried when needed. With a more component-based model, I see how it would be nice to instead manage the Editors as a whole using that map kind of setup, then just store the file handles on each Editor separately, so the scope of the state is unique to each instance. Rather than having a stack of all the leaves for the trees in your forest, you should instead keep the leaves for each tree, on the tree, then you can just manage them by dealing with the trees themselves. See what I mean? Something like that :P

Looks like you finally made it, ... FXMLDR