Un-Constant Element References
Going back on this model, I don't like using static references in this case, at least for ID-based references.

Figured out the bug as to why the tab and or editor were each different colors, it was because the editor's background wasn't properly being assigned to the `--workspace-background` custom property, which it used to use the last I knew of.

Looking into web frameworks for the front end, I'm trying to get used to them, but they just don't feel natural to me still. I think I don't like being locked into one kind of mindset that you have to employ to use the framework (changing your whole codebase over to it, vs starting fresh with a new project?), and I'm still not a big proponent for web tooling. I don't want to use it as much as possible, I think maybe just because I like only building what you absolutely need to have.

I think it's hard to figure all of this out when you aren't quite in your coding-game.

And while this project is fully typed-out and everything, it's still stressing me out how much complexity it still holds. I think I have yet to properly employ and understand a proper component model for a site. That could be my limitations at the moment too, maybe hence why frameworks don't seem to work for me yet.