Card Component Syntax
Properly moved over the rest of the references to the base card component over to the new syntax that I meant to do probably two years ago now, lol.

The card component was originally made with just plain loops and scripts to add the functionality for each instance, and that involved selecting each with a class/attribute query selector, rather than using a full custom element Web Component, which is what I ended up finding out how to utilize later on.

This essentially brings more of the capabilities that are available to you when using custom elements.