Nesting: Demo, Part One
This works amazingly great!!! Found out that hovering on the selector in VS Code shows what kind of HTML it looks like for the element you are selecting. This is super helpful to check that the selector you made is accurate to what you are trying to locate.

I was trying to do the whole while in one-go, but I decided to hold off on that, because there's a TON of nesting to be done inside of the `<main>` styles, and I don't want to break something because I missed an inconsistency in the starting styles in the first part. Going to do this in stages because of that, want to ensure that every change I make doesn't affect the current styling of the site!

Can't remember if I added these as to what I've been listening to lately; last night and tonight :)
Groove has been my new listen recently, it's been a nice creative reset, which I've been hoping to get with my music writing as of the last little bit. Same with progressive black metal, lol. These are two contrasting styles which are working great listening to between, haha.

https://www.youtube.com/watch?v=e6-HossxeWc
https://www.youtube.com/watch?v=HHOn8u-c2wk

https://www.youtube.com/watch?v=rk_BTrL6rwI&list=PL9FMwl2Y6WAdiuvoMqGrVdktDPuFts51Q&index=1
https://www.youtube.com/watch?v=BHDAKPW4lgk&list=PL9FMwl2Y6WAdiuvoMqGrVdktDPuFts51Q&index=2