Bundler-Based File Structures
Thinking about actually trying out React here. That might be too big of a project, but I'm curious of how good it would or wouldn't work for this project. I have to do a few bundler-related things to allow everything to work with JSX though, so that's what I'm doing here. This will allow me to import the files and use them in components with `<img>` tags.