Standalone in macOS Sonoma Safari
The original handling for detecting standalone mode/macOS devices fell out of date because of Safari 17's new standalone installation mode for PWAs in macOS Safari (yay!)

This meant that macOS was detected as being iOS while using the app on macOS in Safari 17. This made it so the widget card title bar styling appeared like the regular non-macOS styling, which isn't intentional. I purposefully designed a style to match the macOS aesthetic real nicely, and that stopped working here.

The fallback here is to instead validate against whether the device is a touch device or not, instead of whether `navigator.standalone` is available.

https://stackoverflow.com/questions/9038625/detect-if-device-is-ios
https://developer.mozilla.org/en-US/docs/Web/API/Navigator/maxTouchPoints