React Setup!
Similar to Dovetail, I'm curious as to how nicely (or not) as to whether this project can be built with React. I think that probably just depends on me knowing how to use hooks and JSX effectively. It probably will work more ergonomically, and it's more maintainable that way too, with actual components and such.

I think part of my push to try and move towards bigger mainstream technologies as of lately is because I don't want to stagnate in my original learned skills, I want to be a better developer, and seeing the great things you can do with modern tooling is really inspiring me. I need to know the initial building blocks to be able to make good use of those though. Say like for Theo's T3 stack, I know how to use TS now, but that's about it. I want to learn about Server Components, and try out SSR and SSG, but I need to know how to use React (or another framework) to be able to accomplish that nicely. I want to try reinventing the wheel less and less, for the sake of my own learning. That was what I've needed to start out I think though, so I'm glad I did that. I think I'm finally turning a corner on accepting using more-advanced and abstracted away tooling though, so I'm glad that I won't be a hindrance for myself in regards to that. I think I've probably been my biggest enemy up until now. But that's also my best friend, because now I feel right at home with the vanilla stack. It just doesn't feel as efficient as I am noticing the other options appear to be out there (like component-based frameworks), so I'd rather take longer learning those than put more time continuing to roll my own stack and try to figure things out from the ground up. I do want to continue to fully understand how the stack I choose works from the ground up (I like doing that, I think it's also my due-diligence to know how it works), but not necessarily need to maintain or make it myself. It's okay to rely on others' work!

This is mostly an experimental demo, things aren't really actually working yet.

It is nice to see that the CSS for the built JSX actually works perfectly fine since it uses the same HTML afterall, so that's a big hope provider for me that it's actually not too hard to migrate to React, because I can leave the CSS as-is initially, and refactor it into the components once I have them built, rather than needing to also move the CSS when I design the new components themselves.

I'm still curious about a Web Component-backed framework that uses JSX for `document.createElement()` calls. Found some recommendations from others, seems like SolidJS could be a nice example/alternative to get close to that kind of thing.