Syntax Highlighting Toggle Fix
Wow! This has been plaguing me for a while, and I couldn't figure out what the issue was. Turns out I just forgot to update the Editor querySelector logic to reflect the addition of the `ste-editor` component, since behorehand it was only set up/needed to handle finding `num-text` elements.