Typing for Settings Properties
STE had it's own layer to abstract away use of Local Storage, but it wasn't typesafe, so I was casting things all over the place to ensure the values were right, and even then, they weren't being validated at runtime, so they may have had a few instances where they were different than what the assertions described them as.

This setup also isn't ideal in the longrun, but it's much better than the other setup. I think doing this kind of thing is what STE needs, just better practices for existing structures it's built out of. But hey, now I get autocomplete for setting values, and I get type inferences for when I assign the wrong types to them! This isn't new to me, but this was much more hectic to manage in STE's type stack previously, since I didn't code it in a TS-friendly way, because I wasn't using that yet.