macOS Card Controls Removal
This was a really nice feature, and I'm glad I added it! It turned out really realistic, and it worked nicely too. But I want to experiment with STE's design a bit more, maybe simplify it a bit more, bring it back to it's roots, or just modernize it a little. And, maybe, just maybe, add a light mode too (oohh noooo). I tried doing that a little ways back, but that was a rework that I never brought forward. I'll go back to that to look for inspiration for it.

But yeah, I'm not removing this because I don't like it, or it didn't work good, but because I want to try new things again, and I don't want it to hinder my design decisions later on, since I want to have new layouts and styles, and I think this time around I want to try using the same designs for all platforms.

Here's to new adventures!