Editor Referencing + File Handle Pairing
Breaking down the use of the global `STE` object, little by little. Part of that involves reorganizing use of the `STE.query()` function, which is what has been used to work and manage state for the Editors. Now that they have their own component models, this can actually be nicely restructured so they can hold their own state, and also hence be passed by reference rather than by ID queries. Since I didn't have a class that holds data for each Editor instance, I would query the document to get the data out of the Editor's individual sections, and get the data that way. I didn't yet understand the concept of using something like a class, where it contains it's own data, and you use it itself to manipulate that data. I only knew the concept of querying the DOM for data, so that's how I built it the first time. Then I looked into getters and setters, and that vastly helped out too. I think I just didn't come back and adjust this setup since initially implementing it, so it's been more complex than it's needed to be, for a long while now. Good to be finally coming around to move things forward again :D

I think I'm still in the stage of modernizing the current STE, I don't feel like I'm 'moving forward' on it yet, but rather for preparing to be able to do that. This is no doubt making STE better with new changes, but it's mostly reworking the original chaos that built up over the years of developing things.

I've also thought about and tried to just reimplement the whole thing from scratch again, since I'm probably going to have a new codebase down the line anyways, but yet again, I learn to remember that you can't remember every minute detail of how things are implemented, you might as well migrate everything in stages, since you can't forget anything that way, because you do it in levels of adjustments, rather than full duplications.