STE.query() Sunsetting!
Finally breaking apart how this works! It was very stacked up and weird, looking back on it. But it's all based on experience, or rather my lack there of, always lots to learn from.

Very mind opening experience to hear the origins behind Devin Townsend's album Infinity, now I have a mix of feelings about the album. I love the music a lot, and still do, but it's a similar sad scenario to City, where it was heavily based on strife, or maybe rather it's relation to it (he mentions how it's a Ying-Yang based approach, I think the other was City? This is the peaceful side to the two). He's very inspiring in how he reflects on his life and his outlooks over time. Where he is now is so much different than back then.

Devin Townsend Podcast is great.