Editor State Properties
Moving the previous state handling away from `data-` attribute specific logic, in favor of using class fields instead. This still inherently relies on the data properties, but now they are accessed and updated using class fields rather than manually updating the attributes on the elements themselves. So then next I will just need to use an internal class field to hold the state of the property, and only use the `data-` attributes for styling and other things like that. It shouldn't be used to specifically hold state for the component. It may be a different story for another component, I think data attributes work really well for that! But for this, I don't want to duplicate the logic of working with the attributes, I'd rather everything be more property-based, and explicitly/securedly type-safe as well. The hinting here is much better too, since it's something that's physically on the Editor instance that you can access, rather than needing to remember to access the attributes from the tab element.

But who knows, maybe using `data-` attributes internally behind a getter-setter pair is actually kind of neat! It's one less layer of state management to handle, and it's nice and clean from the usability perspective. It also stays accurate to if you were to manually update the attribute outside of STE's logic, say with your own scripts (for some reason), or with DevTools. I guess the only issue is that I haven't hooked this up to `attributeChangedCallback()`, so it's not completely two-directional yet, but it would work nicely if that were the case. Also, these attributes should probably just be on the `<ste-editor>` element instead, rather than the tab element. I put it there because that's the only place that state was really needed, for styling essentially. So if I could store it there, use it for styling, and just query that spot to get the state for the component essentially, then that's what I went with. It still makes sense, I think the hard part was just that I manually hooked all of them together, rather than using component-based handling to logically hook everything together for me. That's essentially what I'm trying to alleviate here! :)

Also, if I'm going to use this path going forward, I probably should just use the `dataset` property, since that's easier to work with booleans. *Edit: They don't! I was misnomered about that. It still shows as an empty string if you set a dataset value as a boolean.

```html
<div id="gg" data-fart-breath></div>
<script>
console.log(gg.dataset.fartBreath);
</script>
```
https://developer.mozilla.org/en-US/docs/Learn/HTML/Howto/Use_data_attributes
https://stackoverflow.com/questions/16864999/can-i-use-html5-data-attributes-as-boolean-attributes
https://developer.mozilla.org/en-US/docs/Glossary/Boolean/HTML