Static Editor Method Removal
Discussed in the previous commit, now the best way to work with Editor instances are with the static `Editor.query()` function, then using the instance methods on each Editor itself to manage the calling of it's state hooks. That sounded much more fancy and verbose than waht this does. What?