setView Async Animation Handling
Same course of action with this one too, now you can await until the end of the animation is complete. I think this is how I can help refactor the animation duration handling, which is currently all over the place and confusing. I think part of the trouble was not being able to account for the animation's run time, when running code related to it in JS. Say when you update the size of an element, and it has a resize transition, the function call needed to wait for that transition to complete, before calling the next action. I was doing that with timeouts previously, which was complex to manage because it's not as easy to track what parts of the function are running when. At least, I think it's easier to understand in a Promise-based handling setup.