Omnibox Button Scaling Removal
This didn't work on Retina displays, and I don't know how to account for those, since you can't reliably determine what the browser's initial zoom level is when the page starts. If we knew that, then we could deduce what the current one is, and account for that, but this isn't something heavy enough I want to account for, since it's such a small feature.

This removal simply means my custom Omnibox buttons will now scale up and down when zooming in and out of the page. This is not as usable in terms of the app's appearance when installed, with WCO enabled, but I also don't want to make it extra hacky just to support it. If there is a way to manage this consistently, without too much work, then I would like to try it out. But I'm not sure how zooming should work with WCO in general, I don't think that's something the API completely manages too well, I don't know how it *should* be managed either, since it's just part of the rendering for HTML in that area anyways.

I think I will rework the header a little bit more anyways, so this is another feature that I'm removing, to allow for easily moving on to something else if needed anyways. This was one of those polish kind of features anyways, to complete the header's look and make it consistently responsive.

Just bought Wooden Smoke tonight, been meaning to properly get it for the last few months now :)