4️⃣ 🎂 + Dependency Upgrades
Upgrading to the latest version of things! I'm keeping Better TypeScript at it's current version here because it had some breaking changes later on, and I don't think I've written things to account for them yet. In Dovetail, I also moved away from using Better TypeScript, in favor of the regular types from the standard library.
https://github.com/Offroaders123/Dovetail/commit/e4056db0507fd235afbf38a9c49330bc83b6d10f

Finally getting along towards using trailing newlines at the end of files too haha, going to continue working towards that here as well!

Realized it's time to celebrate STE turning 4! I'm a bit late on this one, but still a cool milestone. Was taking inventory on what all of my projects are at now.

Dovetail - almost 1 year old!
NBTify - about 2 years old
STE - 4 years old
Dark Mode/Simple Sides - nearly 6 years old :O

Been making a lot of progress in learning a more in-depth stack over in Dovetail's dev branch. Diving into using SolidJS, it's outstanding! It's pretty much just what I've been hoping to find while looking for frameworks. I really like the use of JSX/TS (vanilla-esque syntax, in the sense that it works right in your editor, doesn't need language extensions), and even moreso that it compiles down to vanilla DOM calls, and it doesn't use a VDOM. It's really helping me understand how React works at a functional/component/state level, since it's closer to the primitives that I use already. I really like the Accessor/Setter pattern, it makes it explicit as to what is reactive, compared to what is just a plain value. If you want it to be reactive, you have to use one/a combination of those. So it's like React, but where you opt-in to reactivity, instead of it being implicitly reactive, where there you have to opt-out of reactivity depending on what you are trying to do. It's really great that updating your components doesn't rerender the whole DOM tree either, it's like Lit components in that it only patches the parts of the DOM that changed, say like a single text node or something like that. Freaking awesome!

So yeah, going to eventually try bringing SolidJS to STE as well, allowing for a more explicitly-composed component tree. It's been very eye-opening to understand where my own concepts were lining up before. It seems like I've tended to see things from a global state perspective, hence why components have felt counter-intuitive for me to set up. Since SolidJS allows for declaring reactive global state, it makes migrating to it much better I feel, because moving to the component model doesn't mean you have to refactor your HTML and your state at the same time. Instead you can migrate your HTML/DOM calls to JSX components, and still hook those components into the global state that you already had set up. Extra helpful! Learning SolidJS to me feels like a stepping stone towards being more seasoned in working in React, and I'm going to keep going that route because of that. I really like SolidJS in terms of how it works, how you use it, and it's simplicity in general. I think it's what I would pick for my own projects at the moment, since it also falls nicely in line with the TS/lightweight kind of mentality. Most projects are built with React though, so Solid isn't the solution for all cases. For me to get a job somewhere to work on things like that, React is one of the key things that (in my eyes at least) is kind of a given, in that it is definitely complex to work with, but people expect it of you to know how to work with it.

Really looking forward to things to come! Just gotta keep working at it, and improving work on putting myself out there.