JSX Root
"End of Summer" album. I really like how concise that project was.

The bass drum sound of our kit sounds like "Johnny One-Note / The Exciting New Toothpaste from Mars" to me, on my "Lordsmith" album. I didn't know "Hat." then, so it wasn't intentional nor on accident, it was a coincidence? I guess that's the same as being an accident.