STE Object Tree Shaking
Well, 'tree-shaking' in the module/declaration/import sense. It's most likely not changing anything in how it looks for the bundle step. I want to do this because I want to break apart the state of the app properly into where the actual components use them. It's kind of all mixed together in one spot currently.