Component-Based Editing Behavior
I wanted to make this a JSX component instead, but things aren't to the point of being setup to where that's possible yet. Having this less based on querySelector calls is already better though!

Here Come the Builders!

I was going to use `onMount()` instead of `requestAnimationFrame()`, but it's usage inside of `Editor` is outside of the SolidJS component tree, so it won't get cleaned up properly.