Checkbox Component
Very happy to finally have this one set up! Wanted to turn this into a Web Component when I moved to that the previous iteration ago. Glad to finally have this abstracted out. I still have to move these over to using signal-based state rather than element references, but I'm going to leave that migration for that setup instead.