Widget Functional Component
Tried in a demo fork to move over the card setup to a JSX component, but the state handling for it currently doesn't translate over very well, so I'm instead going to simply start by wrapping them with a nice and easy abstraction layer, then modify the backing itself, to eventually be plain functional JSX.

I want to mention that I call things JSX in general, because I don't attach the concepts here to a given framework, if it's React, Solid, or something like that. I think the general way you build them is similar overall, how they run and the details of how things update definitely changes, but the mental model for me is nearly the same. It's using dependency injection where possible.

Evermore - DT