Tools Flattening!
So happy to be moving this forward in this way! It makes so much more sense now.

First time listening to the most recent Puppy album! I remember hearing of it back when it came out, but I think I was listening to other things at the time, so it kind of went to the back of my mind.

https://www.youtube.com/playlist?list=OLAK5uy_kn_-IvNdUp5exf_bcGsDoPuDwyR2n9quI