JSON Formatter State + Refs
Num-Text doesn't support onchange yet, I need to work on that component as a whole. Come to think of it, it may make a whole lot of sense to render it with Solid :D Looks like that may not be as feasible as it sounds though, since making cross-framework components with a framework still leads to it's own challenges. Would one write it as a Web Component or something, then provide wrappers on that element for each of the frameworks on their own? Seems like a possibly nice route to take! I like that concept. Fully cross-compatible component libraries. It could be an Electron of frontend frameworks. Not sure if that would be something anyone would be interested in. Say you would use `num-text` as a Web Component, `num-text/solid` for the Solid component, `num-text/react` for React, and so-on.

Valley Mill

https://github.com/solidjs/solid/discussions/839