Header Components Breakup
Made a standalone Menubar component! This makes the Header component much more manageable to understand. There are multiple other instances where I want to do this still.

Get All You Deserve, Zen next!

Found this a few hours ago, I love this breakdown so much! So glad to have it in a text format to be able to read easily.
https://www.reddit.com/r/DevinTownsend/comments/bggpue/empath_track_by_track_transcript/