Solid Store for Editors
Solid stores are freaking great! Their reactivity is super awesome. It took a bit to figure out how to remove keys from an object defined with a store.

I was going to try moving over to an array of Editors instead, since I thought the object wasn't going to work, but it ended up working thankfully. Looking at it now, I would want to use an array instead anyways, but I'd really like to not change the functionality of the code at all during this migration, only make things more modular and modern.

I think stores could be an insanely great way to work with NBT data in a reactive app, definitely going to have to look into that. Maybe Dovetail's tree view could utilize that.

https://docs.solidjs.com/concepts/stores
https://www.reddit.com/r/solidjs/comments/16ld299/how_do_i_use_setstore_to_delete_an_array_member/
https://github.com/solidjs/solid/discussions/1750
https://github.com/solidjs/solid/discussions/2150
https://raqueebuddinaziz.com/blog/comprehensive-guide-to-solid-js-stores/

Lightnin' Roy, I need to listen to this one more often!!!