Editor Class Feature Flattening
Slowly but surely breaking apart the Editor class into standalone state, component, and rendering logic. Right now it's a complete spaghetti CF mess of parts grossly plucked together. I want to think it's not that bad, but it's really bad 😅

I started working on this the last day or so, I realized these changes can work on their own. I think a nice way to move things forward is to try it in a dev branch, see how your concepts for the migration work/don't work, stash those changes, then go back to the more stable branch, and commit them little by little, adjusting things to something maintainable in the long-run, to ensure that you can continue to add the next features as part of the migration.

I'm learning that a lot of my concepts for this were done very nicely, it just wasn't very dedupe/standalone/modular friendly. Essentially right now, my state is tied to my UI, which is also hooked to the reactive hooks/method calls. So it's hard to split things apart, because some code is state-reactive, while the next line might be adjusting the UI state, from those changes. So there isn't a single source of truth for the values. I think this is what I ran into with the slow down of development for STE, adding these new features, I didn't know how to extract the state out of the UI, and I was getting wrapped up in trying to handle that, and that's what prevented things from moving forward. I think maybe other people might have started fresh on a project like this, I was going to do that as well. I kind of didn't want to though, since the functionality I have here already works very well, it's just not structured/setup nicely. Aah! That's the word I was looking for, coupling. The state is tied to the UI, and the elements themselves. It's very high in coupling.

I was also going to mention in the last commits with CSS changes, that I'm really glad that I organized my styles like that, it made it very easy to split things up into modules, worked great! So the only holdup with that was that the elements themselves needed to be more easily representable with functional components, which meant first refactoring how their state is passed around. That's why there's so much to this rebuild, because it's all twisted around like a bunch of computer cables. And yeah, I didn't really want to do this migration, because I could be learning how to use the frameworks instead, but I think it might be the other way around actually. This is the kind of thing you want to know how to handle when using frameworks, how do you move existing projects over to your new framework. If you can't mangle the state into the more streamlined layout, then how are you going to do that in a different project? I have learned so much in refactoring projects, be it a frontend one like this, or a JS project, a TS one, anything. Maybe next I can look into how to refactor an OOP project to FP (I kind of did that with NBTify, it was definitely a headache too), and the other way around too.

At first I thought this project hasn't made any progress since doing all of this future-proof work, but I think it's the complete opposite, while it may not seem like it. Once I have all of this straightened out on it's own, then I can make actual changes to the codebase safely, with a nice and tidy codebase, rather than a hectic one.

I very, very, heavily much need to update the README for this project, it is definitely not vanilla anymore, probably not even close. Amazing how much things can change over time! Let alone just the progression, but the goals as well. Now I really like frameworks. I think it's because I see what plain projects are great at, and now I see where their limits might be.

https://www.youtube.com/watch?v=CwRF8alr6vo